+++ .gitignore
@@ -2,7 +2,7 @@
 venv
 db.sqlite3
 agency_models
-__pycache_
+__pycache__
 # Distribution / packaging
 bin/
 build/
