+++ kinopoisk_app/serializers/UserSerializer.py
@@ -1,0 +1,13 @@
+from rest_framework import serializers
+from kinopoisk_app.models.UserModel import User
+
+
+class UserSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = User
+        fields = [
+            'id',
+            'full_name',
+            'created_at',
+            'favorite_genres',
+        ]
