+++ kinoposik/settings.py
@@ -78,9 +78,14 @@ WSGI_APPLICATION = 'kinoposik.wsgi.application'
 # https://docs.djangoproject.com/en/4.0/ref/settings/#databases
 
 DATABASES = {
+
     'default': {
-        'ENGINE': 'django.db.backends.sqlite3',
-        'NAME': BASE_DIR / 'db.sqlite3',
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+
+        'NAME': 'kinopoisk',
+        'HOST': 'localhost',
+        'PORT': '5432',
+
     }
 }
